backend/requirements.txt
asttokens==2.0.4
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
decorator==5.1.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
gitdb==4.0.7
GitPython==3.1.27
greenlet==1.1.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
networkx==2.5
numpy==1.21.2
pandas==1.2.4
Pillow==8.3.1
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.20
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
